refactor: common 모듈 분리
