+++ kitchenpos-common/src/main/java/kitchenpos/common/exception/NotFoundException.java
