+++ kitchenpos-common/src/test/java/kitchenpos/domain/NameTest.java
@@ -1,7 +1,8 @@
-package kitchenpos.common.domain;
+package kitchenpos.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import kitchenpos.common.domain.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
