+++ kitchenpos-common/src/test/java/kitchenpos/domain/PriceTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.common.domain;
+package kitchenpos.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
+import kitchenpos.common.domain.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
