+++ src/main/java/kitchenpos/order/dto/OrderLineItemResponse.java
@@ -1,7 +1,6 @@
 package kitchenpos.order.dto;
 
 import kitchenpos.common.domain.Quantity;
-import kitchenpos.menu.domain.Menu;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderMenu;
