+++ src/main/java/kitchenpos/ordertable/application/TableService.java
@@ -2,9 +2,6 @@ package kitchenpos.ordertable.application;
 
 import java.util.stream.Collectors;
 import kitchenpos.common.exception.NotFoundException;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.order.validator.OrderValidator;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.dto.OrderTableRequest;
 import kitchenpos.ordertable.dto.OrderTableResponse;
